+++ src/rpc/misc.cpp
@@ -6,6 +6,7 @@
 #include "base58.h"
 #include "chain.h"
 #include "clientversion.h"
+#include "core_io.h"
 #include "init.h"
 #include "validation.h"
 #include "httpserver.h"
