+++ src/test/rpc_tests.cpp
@@ -6,6 +6,7 @@
 #include "rpc/client.h"
 
 #include "base58.h"
+#include "core_io.h"
 #include "netbase.h"
 
 #include "test/test_bitcoin.h"
